Refactored ChannelInformation to ChannelInfo
